requirements.txt
tensorflow>=2.10.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
Pillow>=8.0.0
pathlib2>=2.3.0
streamlit>=1.28.0
plotly>=5.15.0
pandas>=1.5.0
openpyxl>=3.0.0